[fix] Update NaverOcrRequest timestamp type to long and adjust OcrService for improved timestamp handling

+++ src/main/java/kr/co/ksgk/ims/domain/ocr/dto/request/NaverOcrRequest.java
@@ -8,7 +8,7 @@ import java.util.List;
 public record NaverOcrRequest(
         String version,
         String requestId,
-        Integer timestamp,
+        long timestamp,
         List<ImageInfo> images
 ) {
     @Builder
